fix build issue and warnings in 5.1 plugin (#68)
error: non-constant-expression cannot be narrowed from type 'int32' (aka 'int') to 'float' in initializer list [-Wc++11-narrowing]
fix plugin build warnings monolithic

+++ InworldAI/Source/InworldAIIntegration/Private/InworldPlayerAudioCaptureComponent.cpp
@@ -58,8 +58,8 @@ void ConvertAudioToInworldFormat(const float* AudioData, int32 NumFrames, int32 
         Audio::FResamplingParameters ResamplerParams = {
             Audio::EResamplingMethod::Linear,
             gNumChannels,
-            SampleRate,
-            gSamplesPerSec,
+            (float)SampleRate,
+            (float)gSamplesPerSec,
             InputBuffer
         };
